Round angles to int.

+++ trunk/zquake/cl_tent.c
@@ -459,12 +459,12 @@ void CL_UpdateBeams (void)
 		}
 		else
 		{
-			yaw = Q_atan2(dist[1], dist[0]) * 180 / M_PI;
+			yaw = (int)(Q_atan2(dist[1], dist[0]) * 180 / M_PI);
 			if (yaw < 0)
 				yaw += 360;
 	
 			forward = Q_sqrt (dist[0]*dist[0] + dist[1]*dist[1]);
-			pitch = Q_atan2(dist[2], forward) * 180 / M_PI;
+			pitch = (int)(Q_atan2(dist[2], forward) * 180 / M_PI);
 			if (pitch < 0)
 				pitch += 360;
 		}
